Fix apdu decoding (#147)
This change sets a clear interface for the Generic Message
class. Several attributes have been renamed and protected
from unwanted meddling. 

Code reuse has been improved by moving most of the logic
upwards, such as serialization and common type of decoders
TLV and CBOR. Due to that, the positioning decoding changes
have been tackled with this commit as well.

Tests for asserting the operation of the decoders have been
added to this change and are set to run on each change.

Closes #124, Closes #86, Issue #30, Issue #33