Diagnostic packets: definition and visualization (#182)
This commit updates the cddl definitions for the
247 diagnostic messages and adds a few methods
to allow visualization on top of 3rd party software.

This change contains a change to the mqtt viewer
that allows keys to be flattened. It also adds support
to generate an error message that contains the basic
information regarding the sender of the message 
that caused it.